% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_status.R
\name{addStatusByCutoff}
\alias{addStatusByCutoff}
\title{Add status by cutoff}
\usage{
addStatusByCutoff(
  resDf,
  splitUpDown,
  splitCol = "logFc",
  splitMidPoint = 0,
  metricCol = "logFc",
  metricCutoff = 1,
  sigCol = "pAdj",
  sigCutoff = 0.05,
  noChangeLabel = "No change",
  statuCol = "status",
  upLabel = "Up",
  downLabel = "Down",
  sigLabel = "Significant"
)
}
\arguments{
\item{resDf}{The data frame to annotate.}

\item{splitUpDown}{Split annotation into up and down features? If TRUE, splitCol and splitMidPoint should be provided.}

\item{splitCol}{If splitUpDown = TRUE, the column used to split features into up and down.}

\item{splitMidPoint}{If splitUpDown = TRUE, the splitCol value considered as midpoint.}

\item{metricCol}{Column containing a metric to stablish the cutoff.}

\item{metricCutoff}{The metric cutoff value. Rows with an absolute metricCol value above it will be annotated.}

\item{sigCol}{Column containing a significance value to stablish the cutoff.}

\item{sigCutoff}{The significance cutoff value. Rows with an sigCol value below it will be annotated.}

\item{noChangeLabel}{Label to be used in not annotated rows.}

\item{statuCol}{Name of the status column.}

\item{upLabel}{Label for up-regulated features.}

\item{downLabel}{Label for down-regulated features.}

\item{sigLabel}{Label for significant features. Only used if splitUpDown = FALSE.}
}
\value{
The annotated data frame.
}
\description{
Adds a "status" column to a data frame using a specific cutoff.
}
