% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_status.R
\name{addStatusByRank}
\alias{addStatusByRank}
\title{Add status by rank}
\usage{
addStatusByRank(
  resDf,
  rankCol,
  topN,
  splitUpDown,
  decreasing = TRUE,
  absolute = TRUE,
  splitCol = "logFc",
  splitMidPoint = 0,
  noChangeLabel = "No change",
  statusCol = "status",
  upLabel = "Up",
  downLabel = "Down",
  sigLabel = "Significant"
)
}
\arguments{
\item{resDf}{The data frame to annotate.}

\item{rankCol}{The column used to order the data frame.}

\item{topN}{The top N features to be annotated.}

\item{splitUpDown}{Split annotation into up and down features? If TRUE, splitCol and splitMidPoint should be provided.}

\item{decreasing}{Order rankCol decreasing?}

\item{absolute}{Use absolute rankCol values?}

\item{splitCol}{If splitUpDown = TRUE, the column used to split features into up and down.}

\item{splitMidPoint}{If splitUpDown = TRUE, the splitCol value considered as midpoint.}

\item{noChangeLabel}{Label to be used in not annotated rows.}

\item{statusCol}{Name of the status column.}

\item{upLabel}{Label for up-regulated features.}

\item{downLabel}{Label for down-regulated features.}

\item{sigLabel}{Label for significant features. Only used if splitUpDown = FALSE.}
}
\value{
The annotated data frame.
}
\description{
Order the input data frame using a specific column and annotates
the top N features.
}
