% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeMatrix.R
\name{analyzeMatrix}
\alias{analyzeMatrix}
\title{Default analysis of normalized matrix in SummarizedExperiment}
\usage{
analyzeMatrix(se, groupCol, funCategories, useLimma = TRUE, ...)
}
\arguments{
\item{se}{Summarized Experiment.}

\item{groupCol}{Column from \link[SummarizedExperiment]{colData} used to group samples.}

\item{funCategories}{Functional categories to use in \link[clusterProfiler]{enricher} and \link[clusterProfiler]{GSEA}.}

\item{useLimma}{use \link[biokit]{autoLimma}? If not, uses \link[package]{autoMatrixPairwise}.}

\item{...}{Rest of arguments passed to \link[biokit]{defaultFromTable}.}
}
\value{
A list with results
}
\description{
Default analysis of normalized matrix in SummarizedExperiment
}
