% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDeTable.R
\name{analyzeMultiComp}
\alias{analyzeMultiComp}
\title{Default analysis from multi comparison data frame.}
\usage{
analyzeMultiComp(
  deTable,
  funCategories,
  useCutoff = TRUE,
  fcCutoff = 1,
  pValueCutoff = 0.05,
  n = 50,
  sortCol = "pAdj"
)
}
\arguments{
\item{deTable}{Data frame with stat results.}

\item{funCategories}{Functional categories to use in \link[clusterProfiler]{enricher} and \link[clusterProfiler]{GSEA}.}

\item{useCutoff}{Use cutoff to annotate up and down features?}

\item{fcCutoff}{Fc cutoff. Passed to \link[biokit]{annotateByCutoff}.}

\item{pValueCutoff}{P value cutoff. Passed to \link[biokit]{annotateByCutoff}.}

\item{n}{N features to annotate. Passed to \link[biokit]{annotateTopN}.}

\item{sortCol}{Column used to sort data.  Passed to \link[biokit]{annotateTopN}.}
}
\value{
Data frame containing results of annotating deTable, ORA and GSEA and several default plots.
}
\description{
Performs the "default" analysis given a tidy data frame with the statistical results of several comparisons.
}
