% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateStats.R
\name{annotateByCutoff}
\alias{annotateByCutoff}
\title{Add status column to data frame using cutoff}
\usage{
annotateByCutoff(
  x,
  metricCutoff = 1,
  sigCutoff = 0.05,
  splitUpDown = TRUE,
  metricCol = "logFc",
  sigCol = "pAdj",
  noChangeLabel = "No change",
  annotationCol = "status",
  upLabel = "Up",
  downLabel = "Down",
  sigLabel = "Significant"
)
}
\arguments{
\item{x}{Data frame to annotate.}

\item{metricCutoff}{Metric cutoff. If set to \code{NULL}, only the significance cutoff is used.}

\item{sigCutoff}{Significance cutoff. If set to \code{NULL}, only the metric cutoff is used.}

\item{splitUpDown}{Split generated label into up and down features? If TRUE, uses \code{-metricCutoff} to select down features.}

\item{metricCol}{Column used to filter by metric.}

\item{sigCol}{Column used to filter by significance.}

\item{noChangeLabel}{Label used for not-changing features.}

\item{annotationCol}{Name of the new column containing labels.}

\item{upLabel}{Label for up-regulated features. Only used if \code{splitUpDown = TRUE}.}

\item{downLabel}{Label for down-regulated features. Only used if \code{splitUpDown = TRUE}.}

\item{sigLabel}{Label for significantly altered features. Only used if \code{splitUpDown = FALSE}.}
}
\value{
Annotated data frame.
}
\description{
Uses a significance and metric cutoff to annotate a
data frame containing the results of a stat analysis.
}
