% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{annotateMultiComparison}
\alias{annotateMultiComparison}
\title{Annotate multiple comparisons data frame}
\usage{
annotateMultiComparison(x, useCutoff, compCol = "comparison", ...)
}
\arguments{
\item{x}{Multi-comparison data frame to annotate.}

\item{useCutoff}{Wether to use \link[biokit]{annotateByCutoff} or \link[biokit]{annotateTopN}.}

\item{compCol}{Column with the comparison id.}

\item{...}{Rest of arguments passed to \link[biokit]{annotateByCutoff} or \link[biokit]{annotateTopN}.}
}
\value{
Annotated data frame.
}
\description{
Uses \link[biokit]{annotateByCutoff} or \link[biokit]{annotateTopN} to annotate a
data frame by sepparating data for each comparison.
}
