% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer_functions.R
\name{annotatePeaks.pl}
\alias{annotatePeaks.pl}
\title{annotatePeaks.pl}
\usage{
annotatePeaks.pl(bedfile, genome = "mm10", specific_motifs = NULL,
  mscore = F, mbed = NULL, center = NULL, annStats = NULL, outputfile,
  run_as_sys = F, win_sys = F)
}
\arguments{
\item{bedfile}{The full path to the bed file}

\item{genome}{the genome to use; "mm10" is default.}

\item{specific_motifs}{Search for specific motifs? Takes a vector; adds the "-m" flag.}

\item{mscore}{Requires specific_motifs; returns log-odds score. logical.}

\item{mbed}{return specific locations of the motifs found.  argument takes the string of the full path to the output bed file.}

\item{center}{Center on a motif? argument takes the string of the full path to the output bed file.}

\item{annStats}{Returns a detailed set of annotation statistics. argument takes the string of the full path to the output txt file.}

\item{outputfile}{Full path to the annotated output text file.}

\item{run_as_sys}{run this command through system? False by default, returning the command string.}

\item{win_sys}{Set to False; this enables the system command to work properly on windows.}
}
\value{
By default, this returns a string of the compiled command.  If run_as_sys == T, this will run the command as well.
}
\description{
builds a command to run Homer's annotatePeaks.pl with a set of commonly used flags.
}
\examples{
# annotatePeaks.pl(bedfile = "example.bed",genome = "mm10",outputfile = "annotated_output.txt")
}
