% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_biokit_analysis.R
\name{autoBiokitAnalysis}
\alias{autoBiokitAnalysis}
\title{Default automatic biokit analysis}
\usage{
autoBiokitAnalysis(
  mat,
  sampInfo,
  groupCol,
  funCatList,
  statMode,
  annotationMode,
  filterByExpr,
  splitUpDown = TRUE,
  splitCol = "logFc",
  splitMidPoint = 0,
  sigCutoff = 0.05,
  metricCutoff = 1,
  rankCol = "logFc",
  topN = 50,
  decreasing = TRUE,
  absolute = TRUE,
  minFunCatSize = 1,
  maxFunCatSize = Inf,
  gseaRankCol = "logFc",
  gseaNPerm = 10000
)
}
\arguments{
\item{mat}{The input matrix containing normalized data or counts}

\item{sampInfo}{The sample information data frame.}

\item{groupCol}{The column containing the grouping variable}

\item{funCatList}{A list of character vectors with the functional categories to analyze.}

\item{statMode}{Type of statistical analysis to perform the row-wise comparison.
Allowed values are: "edgeR", "limma" "tTest",
which will apply \link[biokit]{autoEdgerComparison}, \link[biokit]{autoLimmaComparison} and
\link[biokit]{autoPairwiseMatrixTest}, respectively.}

\item{annotationMode}{Approach to annotate differential features. Allowed values are: "byCutoff or "byRank",
which will apply \link[biokit]{addStatusByCutoff} or \link[biokit]{addStatusByRank}.}

\item{filterByExpr}{If "edgeR" is selected, use \link[edgeR]{filterByExpr}?}

\item{splitUpDown}{Split annotation into up and down features? If TRUE, splitCol and splitMidPoint should be provided.}

\item{splitCol}{If splitUpDown is set to TRUE, the column used to split features into up and down.}

\item{splitMidPoint}{If splitUpDown is set to TRUE, the splitCol value considered as midpoint.}

\item{sigCutoff}{If annotation mode is "byCutoff", the adjusted p value cutoff to use.}

\item{metricCutoff}{If annotation mode is "byCutoff", the absolute log FC cutoff to use.}

\item{rankCol}{If annotation mode is "byRank", the column used to rank results.}

\item{topN}{If annotation mode is "byRank", the top N features to annotate.}

\item{decreasing}{If annotation mode is "byRank", order rankCol decreasing?}

\item{absolute}{If annotation mode is "byRank", use absolute rankCol values?}

\item{minFunCatSize}{Minimal size of a functional category to test. All functional categories below the threshold are excluded.}

\item{maxFunCatSize}{Minimal size of a functional category to test. All functional categories aboce the threshold are excluded.}

\item{gseaRankCol}{Column used to create pre-ranked gsea list.}

\item{gseaNPerm}{Number of permutations to perform in the pre-ranked gsea analysis.}
}
\value{
A list containing three data frames: Statistical results, over-representation analysis results and gene set enrichment analysis results.
}
\description{
This wrapper function performs the  "default" analyses of the package.
}
