% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgerUtils.R
\name{autoEdgeR}
\alias{autoEdgeR}
\title{Automatic edgeR analysis from SummarizedExperiment with counts}
\usage{
autoEdgeR(se, groupColumn, minCounts = 0, useFilterByExpr = TRUE, ...)
}
\arguments{
\item{se}{The SummarizedExperiment object to analyze.}

\item{groupColumn}{The id of the \link[SummarizedExperiment]{colData} column that is used to create the pairwise contrasts.}

\item{minCounts}{The minimum number of counts per feature that will be used to filter the count matrix.}

\item{useFilterByExpr}{Wether to use or not the filter obtained from applying \link[edgeR]{filterByExpr}.}

\item{...}{Arguments passed to the \link[biokit]{edgeRFromContrasts} function.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Uses a count matrix stored in a SummarizedExperiment object to
automatically perform all the possible pairwise comparisons
between the groups defined in a \link[SummarizedExperiment]{colData} column.
}
