% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_pairwise_comparison.R
\name{autoEdgerComparison}
\alias{autoEdgerComparison}
\title{Automatic edgeR analysis from counts}
\usage{
autoEdgerComparison(
  counts,
  sampInfo,
  groupCol,
  removeColName = TRUE,
  filterByExpr = FALSE,
  ...
)
}
\arguments{
\item{counts}{The count matrix to analyze.}

\item{sampInfo}{sampInfo The sample information data frame.}

\item{groupCol}{groupCol The column containing the grouping variable}

\item{removeColName}{Remove column name from design matrix?}

\item{filterByExpr}{Apply \link[edgeR]{filterByExpr}?}

\item{...}{Other arguments passed to \link[biokit]{edgerDfFromContrasts}.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Perform the default edgeR analysis given a count matrix, a sample
information data frame and a grouping column.
}
