% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{autoLimma}
\alias{autoLimma}
\title{Automatic limma analysis from SummarizedExperiment}
\usage{
autoLimma(
  se,
  groupColumn,
  compName = "comparison",
  featName = "feature",
  fcName = "logFc",
  pName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{se}{The SummarizedExperiment object to analyze.}

\item{groupColumn}{The id of the \code{colData()} column that is used to}

\item{compName}{Name for the column that indicates the pairwise comparison in the tidy data frame.}

\item{featName}{Name for the column that indicates the analyzed row feature the tidy data frame.}

\item{fcName}{Name for the resulting fold change column.}

\item{pName}{Name for the resulting p value column.}

\item{pAdjName}{Name for the resulting adjusted p value column.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Uses data stored in a SummarizedExperiment object to
automatically perform all the possible pairwise comparisons
between the groups defined in a \code{colData()} column.
}
