% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_pairwise_comparison.R
\name{autoLimmaComparison}
\alias{autoLimmaComparison}
\title{Automatic limma analysis}
\usage{
autoLimmaComparison(mat, sampInfo, groupCol, removeColName = TRUE, ...)
}
\arguments{
\item{mat}{Matrix with data to analyse.}

\item{sampInfo}{sampInfo The sample information data frame.}

\item{groupCol}{groupCol The column containing the grouping variable}

\item{removeColName}{Remove column name from design matrix?}

\item{...}{Other arguments passed to \link[biokit]{limmaDfFromContrasts}.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Perform the default limma analysis given a data matrix, a sample
information data frame and a grouping column.
}
