% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixStats.R
\name{autoMatrixPairwise}
\alias{autoMatrixPairwise}
\title{Automatic pairwise comparison of SummarizedExperiment matrix}
\usage{
autoMatrixPairwise(se, groupCol, ...)
}
\arguments{
\item{se}{The SummarizedExperiment object to analyze.}

\item{groupCol}{The id of the \link[SummarizedExperiment]{colData} column that is used to create the pairwise contrasts.}

\item{...}{Rest of arguments passed to \link[biokit]{matrixTestFromContrasts}.}
}
\value{
A tidy data frame with results.
}
\description{
Uses a matrix stored in a SummarizedExperiment object to
automatically perform all the possible pairwise comparisons
between the groups defined in a \link[SummarizedExperiment]{colData} column.
}
