% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pairwise_comparison.R
\name{autoPairwiseMatrixTest}
\alias{autoPairwiseMatrixTest}
\title{Automatic pairwise contrast}
\usage{
autoPairwiseMatrixTest(
  mat,
  sampInfo,
  groupCol,
  sep = "-",
  compName = "comparison",
  ...
)
}
\arguments{
\item{mat}{Matrix containing data to be analyzed.}

\item{sampInfo}{Sample information data frame.}

\item{groupCol}{Name of the sampInfo column containing the grouping variable.}

\item{sep}{Separator while creating contrasts.}

\item{compName}{Name of the output column that identifies the comparison.}

\item{...}{Other arguments for \link[biokit]{pairwiseMatrixTest}.}
}
\value{
A data frame containing the results for all the comparisons
}
\description{
Perform the automatic pairwise comparison
of matrix data using a sample information
data frame column.
}
