% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.sort.R
\name{bed.sort}
\alias{bed.sort}
\title{Bed Sort}
\usage{
bed.sort(bed_file_path, output_path)
}
\arguments{
\item{bed_file_path}{path to the unsorted bed file.}

\item{output_path}{path to the sorted bed file, the output.}
}
\value{
returns a bed file.
}
\description{
Takes a bed file, sorts it by chromosome and start, then writes it to the output location.
}
\examples{
# bed.sort(bed_file_path = "R/locations.bed",output_path = "R/locations_sorted.bed")
}
