% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.convert.R
\name{bioKIT.unit.convert}
\alias{bioKIT.unit.convert}
\title{bioKIT.unit.convert}
\usage{
bioKIT.unit.convert(df, value, unit)
}
\arguments{
\item{df}{The dataframe holding the calculations.}

\item{value}{value represents the column with the total volume.  Ex: 1234 might be 1,234 ul.}

\item{unit}{the column name, as a string, that represents the unit of the value variable.}
}
\value{
returns a vector containing strings of the value with the attached unit.
}
\description{
Converts the scale of the unit in a dataframe when calculating volumes / mass.
}
\examples{
# ADD_EXAMPLES_HERE
}
