% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blmCounts}
\alias{blmCounts}
\title{RNA-Seq counts for 1000 random genes from the GSE121659 experiment.}
\format{An object of class \code{data.frame} with 1000 rows and 12 columns.}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE121nnn/GSE121659/suppl/GSE121659_raw_counts_merged.txt.gz}
}
\usage{
blmCounts
}
\description{
RNA-Seq counts for 1000 random genes from the GSE121659 experiment.
}
\keyword{datasets}
