% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blmMat}
\alias{blmMat}
\title{RNA-Seq counts from the GSE121659 experiment.}
\format{An object of class \code{matrix} with 27179 rows and 12 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE121659}
}
\usage{
blmMat
}
\description{
RNA-Seq counts from the GSE121659 experiment.
}
\keyword{datasets}
