% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blmSampInfo}
\alias{blmSampInfo}
\title{Sample information from the {GSE121659} experiment.}
\format{
An object of class \code{data.frame} with 12 rows and 1 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE121659}
}
\usage{
blmSampInfo
}
\description{
Sample information from the {GSE121659} experiment.
}
\keyword{datasets}
