% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_calculate.R
\name{buffer_calculate}
\alias{buffer_calculate}
\title{Calculate Buffers}
\usage{
buffer_calculate(buffer.df, amount_col, amount_needed = 1,
  amount_needed_unit = "mL", buffer.name = NULL, buffer.col = "Buffer")
}
\arguments{
\item{buffer.df}{dataframe containing recipes.}

\item{amount_col}{String identifying the values / volumes}

\item{amount_needed}{how much? ex: 5}

\item{amount_needed_unit}{"ml","Samples", etc.}

\item{buffer.name}{filter for a specific buffer?  Takes a String.}

\item{buffer.col}{Default is "Buffer".  Supply this if you are filtering for a specific buffer.}
}
\value{
RETURN_DESCRIPTION
}
\description{
A simplified form of the other buffer calculation object; used to calculate the needed reagents necessary to make X of a buffer.
}
\examples{
# ADD_EXAMPLES_HERE
}
