% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_risk.R
\name{calculate.relative.risk}
\alias{calculate.relative.risk}
\title{Calculate Relative Risk}
\usage{
calculate.relative.risk(case.positive, case.negative, control.positive,
  control.negative, return_df = F)
}
\arguments{
\item{case.positive}{How many in the case population that are positive for the test; integer.}

\item{case.negative}{How many in the case population that are negative for the test; integer.}

\item{control.positive}{How many in the control population that are positive for the test; integer.}

\item{control.negative}{How many in the control population that are negative for the test; integer.}

\item{return_df}{Return a dataframe instead of a string? logical.}
}
\value{
returns a string or a dataframe.
}
\description{
Calculates the relative risk of two populations and returns either a data frame or a string containing the results.
}
\examples{
# ADD_EXAMPLES_HERE
}
