% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{calculate_growth_rate}
\alias{calculate_growth_rate}
\title{Calculate Growth Rate}
\usage{
calculate_growth_rate(current_number, starting_number, time.hours)
}
\arguments{
\item{current_number}{number of cells counted at the end; integer.}

\item{starting_number}{number of cells seeded. integer.}

\item{time.hours}{number of hours that passed between seeding and collection; integer.}
}
\value{
returns a dataframe.
}
\description{
For cell culture.  Calculates the growth rate and doubling time of cells based on hours passed, seeding number, and current count.
}
\examples{
# ADD_EXAMPLES_HERE

}
