% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reagents.R
\name{calculate_reagents}
\alias{calculate_reagents}
\title{FUNCTION_TITLE}
\usage{
calculate_reagents(buffers.df = NULL, Experiment = NULL, Buffer = NULL,
  amount_per_value = NULL, amount_per_unit = NULL, amount_needed = NULL,
  amount_type = NULL, Experiment_col = 1, Buffer_col = 2,
  Reagent_col = 3, Stock_conc_col = 4, Stock_unit_col = 5,
  Final_conc_col = 6, Final_unit_col = 7)
}
\arguments{
\item{buffers.df}{dataframe containing your buffer recipes}

\item{Experiment}{string to filter by experiment. ex: "PCR"}

\item{Buffer}{string to filter by buffer. ex: "6x loading dye"}

\item{amount_per_value}{a numeric value as part of the base recipe for 1 ml/sample. ex: 5, ex: df$amount.per.ml}

\item{amount_per_unit}{a string as part of the base recipe for 1 ml/sample. ex: "ul, ex: df$amount.unit}

\item{amount_needed}{a numeric value for how much of the buffer is needed.}

\item{amount_type}{the unit, as a string, for the amount needed}

\item{Experiment_col}{number of experiment col, if custom}

\item{Buffer_col}{number of buffer col, if custom}

\item{Reagent_col}{number of reagent col, if custom}

\item{Stock_conc_col}{number of stock concentration col, if custom}

\item{Stock_unit_col}{number of stock unit col, if custom}

\item{Final_conc_col}{number of final concentration col, if custom}

\item{Final_unit_col}{number of final unit col, if custom}
}
\value{
a dataframe with the calculations.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
