% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbdCounts}
\alias{cbdCounts}
\title{RNA-Seq counts for 1000 random genes from the {GSE131565} experiment.}
\format{An object of class \code{matrix} with 1000 rows and 6 columns.}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE131nnn/GSE131565/suppl/GSE131565_raw_counts.txt.gz}
}
\usage{
cbdCounts
}
\description{
RNA-Seq counts for 1000 random genes from the {GSE131565} experiment.
}
\keyword{datasets}
