% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbdMat}
\alias{cbdMat}
\title{RNA-Seq counts from the {GSE131565} experiment.}
\format{An object of class \code{matrix} with 28395 rows and 6 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE131565}
}
\usage{
cbdMat
}
\description{
RNA-Seq counts from the {GSE131565} experiment.
}
\keyword{datasets}
