% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbdSampInfo}
\alias{cbdSampInfo}
\title{Sample information from the {GSE131565} experiment.}
\format{An object of class \code{data.frame} with 6 rows and 1 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE131565}
}
\usage{
cbdSampInfo
}
\description{
Sample information from the {GSE131565} experiment.
}
\keyword{datasets}
