% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_culture_guidelines}
\alias{cell_culture_guidelines}
\title{Cell Culture Guidelines}
\format{A dataframe with 10 rows and 6 columns}
\source{
\url{https://www.thermofisher.com/us/en/home/references/gibco-cell-culture-basics/cell-culture-protocols/cell-culture-useful-numbers.html}
}
\usage{
cell_culture_guidelines
}
\description{
A set of data from Thermo Fisher that provides some basic cell culture guidelines.
}
\keyword{datasets}
