% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contrasts.R
\name{contrastsFromDesign}
\alias{contrastsFromDesign}
\title{Create contrast matrix from design matrix.}
\usage{
contrastsFromDesign(desMat)
}
\arguments{
\item{desMat}{The design matrix.}
}
\value{
A contrast matrix with all the possible pairwise comparisons.
}
\description{
Uses the levels defined in the column names of the design matrix
to create a contrast matrix with \link[limma]{makeContrasts}.
}
