% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cells.R
\name{count_cells}
\alias{count_cells}
\title{count_cells}
\usage{
count_cells(sampleID = NULL, live, dead = 0, squares_counted = 4,
  sample_aliquot_vol = 0, PBS_vol = 0, Trypan_blue_vol = 0,
  original_sample_vol)
}
\arguments{
\item{sampleID}{Provides a label or name for the sample counted.}

\item{live}{The total number of live cells counted.}

\item{dead}{The total number of dead cells counted.}

\item{squares_counted}{How many of the outer corner-squares were counted, Default is 4.}

\item{sample_aliquot_vol}{The ul of sample taken for hemocytometry.}

\item{PBS_vol}{The ul of PBS or other solution that the sample aliquot is diluted with. Used to determine the dilution factor.}

\item{Trypan_blue_vol}{The ul of Trypan Blue, if used. A good starting point is usually 1:1 (sample_aliquot+PBS):Trypan Blue.}

\item{original_sample_vol}{How many ml is the original set of cells suspended in?}
}
\value{
returns a dataframe containing all of the calculations.
}
\description{
count_cells() is a calculator for standard hemocytometry applications.
}
\examples{
# count_cells(sampleID = "A",live = 300,dead = 10,squares_counted = 4,sample_aliquot_vol = 10,PBS_vol = 40,Trypan_blue_vol = 50,original_sample_vol = 6)
}
