% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgerUtils.R
\name{countsToTmm}
\alias{countsToTmm}
\title{Normalize counts with TMM}
\usage{
countsToTmm(x, log = TRUE, prior.count = 3)
}
\arguments{
\item{x}{The matrix of counts.}

\item{log}{Passed to \link[edgeR]{cpm}.}

\item{prior.count}{Passed to \link[edgeR]{cpm}.}
}
\value{
The normalized expression matrix
}
\description{
Normalizes a raw counts matrix into log transformed counts per millions using
the edgeR "Trimmed Mean of M-Values" (TMM) normalization method. For more information
please refer to the edgeR package (https://bioconductor.org/packages/release/bioc/html/edgeR.html).
}
