% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraGsea.R
\name{cpResultsToDf}
\alias{cpResultsToDf}
\title{Transform list of clusterProfiler results into tidy data frame}
\usage{
cpResultsToDf(x, splitName = "comparison")
}
\arguments{
\item{x}{The list of clusterProfiler results}

\item{splitName}{The name for the new id column.}
}
\value{
A data frame with the clusterProfiler results in a tidy format.
}
\description{
Binds the rows of a list of clusterProfiler results annotating the
resulting data frame with the list names.
}
