% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filecsv.R
\name{csvFile}
\alias{csvFile}
\title{FUNCTION_TITLE}
\usage{
csvFile(input, output, session, stringsAsFactors)
}
\arguments{
\item{input}{DESCRIPTION.}

\item{output}{DESCRIPTION.}

\item{session}{DESCRIPTION.}

\item{stringsAsFactors}{treat strings as factors? Logical.}
}
\value{
returns a loaded csv file.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
#UI section
#csvFileInput(id = "csv1",label= "load csv file")
#Server Section
#X <- callModule(csvFile,"csv1",stringsAsFactors = F)

}
