% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filecsv.R
\name{csvFileInput}
\alias{csvFileInput}
\title{csvFileInput}
\usage{
csvFileInput(id, label = "CSV file")
}
\arguments{
\item{id}{the shiny id to be called internally, as a string.}

\item{label}{The label that is displayed in the app.}
}
\value{
Returns a load file button in a shiny app.
}
\description{
shiny UI module for loading csv file.  Needs csvFile() in the server function.
}
\examples{
#UI section
#csvFileInput(id = "csv1",label= "load csv file")
#Server Section
#X <- callModule(csvFile,"csv1",stringsAsFactors = F)

}
