% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultFromMatrix.R
\name{defaultFromCounts}
\alias{defaultFromCounts}
\title{Default analysis of counts in SummarizedExperiment}
\usage{
defaultFromCounts(se, groupCol, funCategories, ...)
}
\arguments{
\item{se}{Summarized Experiment.}

\item{groupCol}{Column from \link[SummarizedExperiment]{colData} used to group samples.}

\item{funCategories}{Functional categories to use in \link[clusterProfiler]{enricher} and \link[clusterProfiler]{GSEA}.}

\item{...}{Rest of arguments passed to \link[biokit]{defaultFromTable}.}
}
\value{
A list with results
}
\description{
Default analysis of counts in SummarizedExperiment
}
