% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultFromTable.R
\name{defaultFromTable}
\alias{defaultFromTable}
\title{Default analysis from comparison table}
\usage{
defaultFromTable(deTable, compCol = "comparison", ...)
}
\arguments{
\item{deTable}{Data frame with stat results.}

\item{compCol}{Column containing the comparison id.}

\item{...}{Rest of arguments passed to \link[biokit]{defaultSingleComparison} or \link[biokit]{defaultMultiComparison}}
}
\value{

}
\description{
Uses a column to identify if it is a single-comparison or a multi-comparison data frame and perform
subsequent analyses.
}
