% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{defaultGseaDotPlot}
\alias{defaultGseaDotPlot}
\title{Create default GSEA dot plot}
\usage{
defaultGseaDotPlot(
  x,
  splitById = FALSE,
  idCol = "comparison",
  upLabel = "Up",
  downLabel = "Down",
  upColor = "red",
  downColor = "blue",
  pCutoff = NULL
)
}
\arguments{
\item{x}{Data frame with results to plot.}

\item{splitById}{Use a column id to separate results?}

\item{idCol}{Name of the column used to separate results. Placed at the X axis instead of Up and down features.}

\item{upLabel}{Label used to indicate status of up regulated functional categories.}

\item{downLabel}{Label used to indicate status of down regulated functional categories.}

\item{upColor}{Color used to indicate status of up regulated functional categories.}

\item{downColor}{Color used to indicate status of down regulated functional categories.}

\item{pCutoff}{Adjusted P Value cutoff used to filter results before plotting.}
}
\value{
A ggplot containing the default gsea dot plot.
}
\description{
Creates a simple representation of the results of \link[clusterProfiler]{GSEA} formatted
with \link[biokit]{cpResultsToDf}. The X axis represents comparison or status of the pathway.
The data frame can be filtered beforehand or setting the pCutoff parameter.
}
