% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{defaultLogFc}
\alias{defaultLogFc}
\title{Log 2 of means ratio}
\usage{
defaultLogFc(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{Numerator vector.}

\item{y}{Denominator vector.}

\item{na.rm}{Remove NAs?}
}
\value{

}
\description{
Log 2 of means ratio
}
