% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultFromTable.R
\name{defaultMultiComparison}
\alias{defaultMultiComparison}
\title{Default analysis from multi comparison data frame.}
\usage{
defaultMultiComparison(
  deTable,
  funCategories,
  useCutoff = TRUE,
  fcCutoff = 1,
  pValueCutoff = 0.05,
  n = 50,
  sortCol = "pAdj",
  oraPCutoff = 0.05,
  gseaPCutoff = 0.05
)
}
\arguments{
\item{deTable}{Data frame with stat results.}

\item{funCategories}{Functional categories to use in \link[clusterProfiler]{enricher} and \link[clusterProfiler]{GSEA}.}

\item{useCutoff}{Use cutoff to annotate up and down features?}

\item{fcCutoff}{Fc cutoff. Passed to \link[biokit]{annotateByCutoff}.}

\item{pValueCutoff}{P value cutoff. Passed to \link[biokit]{annotateByCutoff}.}

\item{n}{N features to annotate. Passed to \link[biokit]{annotateTopN}.}

\item{sortCol}{Column used to sort data.  Passed to \link[biokit]{annotateTopN}.}

\item{oraPCutoff}{P value threshold to plot ORA result.}

\item{gseaPCutoff}{P value threshold to plot GSEA result.}
}
\value{
Data frame containing results of annotating deTable, ORA and GSEA and several default plots.
}
\description{
Performs the "default" analysis given a tidy data frame with the statistical results of several comparisons.
}
