% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{defaultOraPlot}
\alias{defaultOraPlot}
\title{Default Over-representation analysis plot}
\usage{
defaultOraPlot(
  x,
  idColumn = "comparison",
  splitStatus = FALSE,
  facetByStatus = TRUE,
  upLabel = "Up",
  downLabel = "Down",
  idSep = "_",
  upColor = "red",
  downColor = "blue",
  pCutoff = NULL
)
}
\arguments{
\item{x}{Data frame with results to plot.}

\item{idColumn}{Column used as id of the analysed group, will be plotted in the X axis.}

\item{splitStatus}{Split id Column into up and down features?}

\item{facetByStatus}{Facet final plot by status of the analysed features?}

\item{upLabel}{Label used to indicate status of up-regulated features in the idColumn.}

\item{downLabel}{Label used to indicate status of up-regulated features in the idColumn..}

\item{idSep}{Separator used to paste group and status label.}

\item{upColor}{Color used to plot enrichments in up regulated features.}

\item{downColor}{Color used to plot enrichments in down regulated features.}

\item{pCutoff}{Adjusted P Value cutoff used to filter results before plotting.}
}
\value{
A ggplot containing the default ora plot.
}
\description{
Creates a simple representation of the results of \link[clusterProfiler]{enricher} formatted
with \link[biokit]{cpResultsToDf}. The X axis represents column of interest while Y axis
represents the functional category. The presence of a point indicates the enrichment of a
functional category in a group of interest. The data frame can be filtered beforehand or
setting the pCutoff parameter.
}
