% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{defaultPcaPlot}
\alias{defaultPcaPlot}
\title{Default Principal Component Analysis Plot}
\usage{
defaultPcaPlot(x, sampInfoDf, groupCol, showLabel = TRUE)
}
\arguments{
\item{x}{Result obtained with \link[biokit]{pcaToList}.}

\item{sampInfoDf}{Data frame containing sample information. Rownames should match \link[stats]{prcomp} matrix rownames.}

\item{groupCol}{Column of the data frame that indicates group of samples.}

\item{showLabel}{Show row labels?}
}
\value{
A ggplot containing the default pca plot.
}
\description{
Creates a simple representation of the \link[stats]{prcomp} results formatted with the
\link[biokit]{pcaToList} function.
}
