% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{defaultVolcano}
\alias{defaultVolcano}
\title{Default volcano plot}
\usage{
defaultVolcano(
  x,
  fcCol = "logFc",
  pCol = "pAdj",
  statusCol = "status",
  compCol = NULL,
  featCol = "feature",
  upColor = "red",
  noChangeColor = "black",
  downColor = "blue",
  upLabel = "Up",
  noChangeLabel = "No change",
  downLabel = "Down"
)
}
\arguments{
\item{x}{Data frame with data to plot.}

\item{fcCol}{Fold change column.}

\item{pCol}{P value column}

\item{statusCol}{Status column.}

\item{compCol}{Comparison column.}

\item{featCol}{Feature column.}

\item{upColor}{Up reuglated features color.}

\item{noChangeColor}{Not chaning features color.}

\item{downColor}{Down regulated features color.}

\item{upLabel}{Up regulated features label.}

\item{noChangeLabel}{Not chaning features label.}

\item{downLabel}{Down regulated features label.}
}
\value{
A ggplot containing the default volcano plots
}
\description{
Creates a default volcano plot from comparison data frame.
}
