% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design.R
\name{designFromSampInfo}
\alias{designFromSampInfo}
\title{Create design matrix from sample metadata}
\usage{
designFromSampInfo(sampInfo, groupCol, removeColName = TRUE)
}
\arguments{
\item{sampInfo}{The data frame that contains sample metadata.}

\item{groupCol}{The sampInfo column that is used to create the design matrix.}

\item{removeColName}{Remove column name from design matrix?}
}
\value{
The design matrix.
}
\description{
Uses a data frame containing sample metadata to create a design matrix
for subsequent analyses.
}
