% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTools2_wrapper.R
\name{dT2_bamCoverage}
\alias{dT2_bamCoverage}
\title{dT2 bamCoverage}
\usage{
dT2_bamCoverage(bam_file, output_file, outFileFormat = "bigwig",
  scaleFactor = NULL, binSize = 10, region = NULL,
  blacklist_file = NULL, effectiveGenomeSize = NULL,
  normalizeUsing = NULL, smoothLength = NULL, extendReads = NULL,
  processors = "max", ignoreDuplicates = T, run_as_sys = F)
}
\arguments{
\item{bam_file}{DESCRIPTION.}

\item{output_file}{DESCRIPTION.}

\item{outFileFormat}{DESCRIPTION.}

\item{scaleFactor}{DESCRIPTION.}

\item{binSize}{DESCRIPTION.}

\item{region}{DESCRIPTION.}

\item{blacklist_file}{DESCRIPTION.}

\item{effectiveGenomeSize}{DESCRIPTION.}

\item{normalizeUsing}{DESCRIPTION.}

\item{smoothLength}{DESCRIPTION.}

\item{extendReads}{DESCRIPTION.}

\item{processors}{DESCRIPTION.}

\item{ignoreDuplicates}{DESCRIPTION.}

\item{run_as_sys}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Wrapper for deepTools2 bamCoverage; converts a bam file to a bigwig / bedgraph for use in visualization and analysis.
}
\examples{
# ADD_EXAMPLES_HERE
}
