% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTools2_wrapper.R
\name{dT2_computeMatrix}
\alias{dT2_computeMatrix}
\title{dT2 computeMatrix}
\usage{
dT2_computeMatrix(bigwig_files, bed_files, scale_regions = F,
  ref_point = "TSS", bin_size = 10, upstream_bp = 1, downstream_bp = 1,
  missingDataAsZero = T, skipZeros = T, blacklist_file = NULL, output_gz,
  outFileMatrix = NULL, outSortedRegions = NULL, run_as_sys = F)
}
\arguments{
\item{bigwig_files}{vector of bigwig filepaths}

\item{bed_files}{vector of bed filepaths}

\item{scale_regions}{use scale-regions mode?}

\item{ref_point}{if using reference-point mode; "TSS","TES",or "center"}

\item{bin_size}{size of the bins to use}

\item{upstream_bp}{how many bases upstream?}

\item{downstream_bp}{how many bases downstream}

\item{missingDataAsZero}{Treat missing data as zero?}

\item{skipZeros}{skip zeros?}

\item{blacklist_file}{use a blacklist file?}

\item{output_gz}{filepath to the output .gz file}

\item{outFileMatrix}{filepath to the output .tab file for underlying values}

\item{outSortedRegions}{filepath to a bed file for outputting sorted regions.}

\item{run_as_sys}{Return the command as a string or run as system}
}
\value{
RETURN_DESCRIPTION
}
\description{
Wrapper for deepTools2 computeMatrix
}
\examples{
# ADD_EXAMPLES_HERE
}
