% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_pairwise_comparison.R
\name{edgerDfFromContrasts}
\alias{edgerDfFromContrasts}
\title{Obtain edgeR results from counts, design and contrasts matrix}
\usage{
edgerDfFromContrasts(
  counts,
  desMat,
  conMat,
  filterByExpr = FALSE,
  compName = "comparison",
  featName = "feature",
  exprName = "AveExpr",
  fcName = "logFc",
  pName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{counts}{The count matrix to analyze.}

\item{desMat}{Design matrix.}

\item{conMat}{Contrasts matrix.}

\item{filterByExpr}{Apply \link[edgeR]{filterByExpr}?}

\item{compName}{Name for the column that indicates the pairwise comparison in the tidy data frame.}

\item{featName}{Name for the column that indicates the analyzed row feature in the tidy data frame.}

\item{exprName}{Name for the resulting average expresion column.}

\item{fcName}{Name for the resulting fold change column.}

\item{pName}{Name for the resulting p value column.}

\item{pAdjName}{Name for the resulting adjusted p value column.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Applies recommended edgeR steps to obtain differential expression tables from
an initial count matrix object. As in \link[biokit]{limmaDfFromContrasts}, it uses
a design matrix and a contrast matrix as input.
}
