% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{edgeRFromContrasts}
\alias{edgeRFromContrasts}
\title{Obtain edgeR results from object, design and contrasts matrix}
\usage{
edgeRFromContrasts(
  object,
  desMat,
  conMat,
  compName = "comparison",
  featName = "feature",
  exprName = "AveExpr",
  fcName = "logFc",
  pName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{object}{The edgeR object to analyze. Generated with: \link[edgeR]{DGEList}.}

\item{desMat}{Design matrix.}

\item{conMat}{Contrasts matrix.}

\item{compName}{Name for the column that indicates the pairwise comparison in the tidy data frame.}

\item{featName}{Name for the column that indicates the analyzed row feature in the tidy data frame.}

\item{exprName}{Name for the resulting average expresion column.}

\item{fcName}{Name for the resulting fold change column.}

\item{pName}{Name for the resulting p value column.}

\item{pAdjName}{Name for the resulting adjusted p value column.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Applies recommended edgeR steps to obtain differential expression tables from
an initial \link[edgeR]{DGEList} object. As in \link[biokit]{limmaDfFromContrasts}, it uses
a design matrix and a contrast matrix as input.
}
