% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number_predict_confluency.R
\name{estimate_confluency}
\alias{estimate_confluency}
\title{Estimate Confluency}
\usage{
estimate_confluency(cell_number, plate_size = "100mm",
  custom_confluency = NULL)
}
\arguments{
\item{cell_number}{DESCRIPTION.}

\item{plate_size}{DESCRIPTION.}

\item{custom_confluency}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Estimates the confluency of a cell culture dish with a known (or predicted cell number), a specific plate size, and the number of cells that are considered 100% confluent on that dish.
}
\examples{
# ADD_EXAMPLES_HERE
}
