% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_representation_analysis.R
\name{fisherExactTest}
\alias{fisherExactTest}
\title{Measure over-representation using Fisher's Exact Test}
\usage{
fisherExactTest(features, funCat, universe)
}
\arguments{
\item{features}{Input features.}

\item{funCat}{All the features in the functional category.}

\item{universe}{All the universe features.}
}
\value{
The over-representation significance as a P value.
}
\description{
Uses \link[stats]{fisher.test} to calculate the over-representation of
a set of features on a functional category given an "universe" that is used
as statistical background. Based on the Fisher's Exact Test over a contingency
table.
}
\examples{

g <- letters[1:5]
f <- letters[2:6]
f2 <- letters[c(1, 10:15)]
u <- letters

# this should give a low p value (high over-representation)
fisherExactTest(g, f, u)
# this should give a high p value (low over-representation)
fisherExactTest(g, f2, u)

}
