% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{getRankedVector}
\alias{getRankedVector}
\title{Get ranked vector from data frame}
\usage{
getRankedVector(x, metricCol = "logFc", featCol = "feature", decreasing = TRUE)
}
\arguments{
\item{x}{Data frame to process.}

\item{metricCol}{Column to extract as ranked vector.}

\item{featCol}{Column containing the names for the ranked vector.}

\item{decreasing}{Sort vector decreasing?}
}
\value{
A vector in the desired format.
}
\description{
Prepares a vector of the given data frame column, after naming and sorting it.
}
