% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsToFunc.R
\name{getRankedVectorList}
\alias{getRankedVectorList}
\title{Get ranked list of vectors from data frame}
\usage{
getRankedVectorList(x, splitCol = "comparison", ...)
}
\arguments{
\item{x}{Data frame to process.}

\item{splitCol}{Column used to split the data frame.}

\item{...}{Rest of arguments passed to \link[biokit]{getRankedVector}.}
}
\value{
A list of sorted and named vectors.
}
\description{
Uses a column to split a data frame and then creates a list of named vectors using
\link[biokit]{getRankedVector}.
}
