% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraGsea.R
\name{gseaFromList}
\alias{gseaFromList}
\title{Perform Gene Set Enrichment Analysis from list of ranked vectors}
\usage{
gseaFromList(x, funCategories, preSeed = 149, ...)
}
\arguments{
\item{x}{The list of ranked vectors.}

\item{funCategories}{Two-columns data frame containing the functional categories info.
First column should correspond to categories and the second to features.
Passed to \link[clusterProfiler]{GSEA} TERM2GENE argument.}

\item{preSeed}{Seed to be set before running each GSEA analysis. Defaults to 149.}

\item{...}{Rest of arguments passed to \link[clusterProfiler]{GSEA}.}
}
\value{
A list of \link[clusterProfiler]{GSEA} results.
}
\description{
Uses \link[clusterProfiler]{GSEA} to perform the gene set enrichment analysis
over a list of ranked vectors.
}
