% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{gseaFromStats}
\alias{gseaFromStats}
\title{Gene set enrichment analysis from statistical results}
\usage{
gseaFromStats(
  df,
  funCatList,
  rankCol,
  featCol = "feature",
  splitCol = "comparison",
  seed = 149,
  nPerm = 1000,
  ...
)
}
\arguments{
\item{df}{Data frame with results to analyze.}

\item{funCatList}{A list of character vectors with the functional categories to analyze.}

\item{rankCol}{Column used to perform the pre-ranked gsea analysis.}

\item{featCol}{Column used to name the rankCol values.}

\item{seed}{Seed used to avoid reproducibility problems between runs.}

\item{...}{Rest of arguments passed to \link[fgsea]{fgseaSimple}.}
}
\value{
A tidy data frame with the results.
}
\description{
Gene set enrichment analysis from statistical results
}
