% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{gseaPlot}
\alias{gseaPlot}
\title{GSEA plot}
\usage{
gseaPlot(gseaResDf, pCutoff = 0.05)
}
\arguments{
\item{gseaResDf}{The output data frame from \link[biokit]{gseaFromStats}}

\item{pCutoff}{Adjusted P value cutoff to filter features before plotting}
}
\value{
A ggplot2 object
}
\description{
Plots the results of a gsea analysis
}
