% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{heatmapPlot}
\alias{heatmapPlot}
\title{Heatmap plot}
\usage{
heatmapPlot(mat, sampInfo, groupCol, scaleBy = "row", nTop = 100, ...)
}
\arguments{
\item{mat}{The matrix to be plotted}

\item{sampInfo}{The sample information data frame}

\item{groupCol}{The column name of the grouping variable}

\item{scaleBy}{character indicating if the values should be centered and scaled in either the
row direction or the column direction, or none. Corresponding values are "row", "column" and "none"}

\item{nTop}{Number of features to be plotted after ranking them by the standard deviation}

\item{...}{Arguments to be passed to \link[pheatmap]{pheatmap}}
}
\value{
The phetamap plot
}
\description{
Creates a heatmap representation of the more variable features in a matrix
}
