% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanHallmarks}
\alias{humanHallmarks}
\title{Human MSigDb hallmarks}
\format{
An object of class \code{list} of length 50.
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp}
}
\usage{
humanHallmarks
}
\description{
A list of vectors containing the functional categories defined by the MSigDb functional hallmarks.
}
\keyword{datasets}
