% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_PCR.R
\name{in_silico_PCR}
\alias{in_silico_PCR}
\title{in silico PCR}
\usage{
in_silico_PCR(forward_primer, reverse_primer, mRNA_primers = F,
  genome_version = "mm10", max_product_size = 4000,
  min_perfect_match = 15, min_good_match = 15, flip_reverse_primer = F)
}
\arguments{
\item{forward_primer}{string containing the forward primer sequence.}

\item{reverse_primer}{string containing the reverse primer sequence.}

\item{mRNA_primers}{Test the UCSC Genes assembly? False by default, tests the genome assembly.}

\item{genome_version}{mm10 by default, mm9 is also valid.}

\item{max_product_size}{max bp size for the product.  integer.}

\item{min_perfect_match}{minimum perfect match, typically left as is.}

\item{min_good_match}{minimum good match, typically left as is.}

\item{flip_reverse_primer}{flip the reverse primer? False by default.}
}
\value{
returns a string of the predicted amplification region(s).
}
\description{
Builds a request and parses the output of UCSC's in silico PCR.
}
\examples{
#negative result
#in_silico_PCR(forward_primer = "AGCCTCCAAAACCAAGGTGA",reverse_primer = "ATCTGGTCAACCACCTCCTT")
#positive result
#in_silico_PCR(forward_primer = "TTCTCTGTCATGGAACCCTATTC",reverse_primer = "CAGGAGAGAGTAGTCCACAGTA")
}
