% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflect.R
\name{inflect}
\alias{inflect}
\title{Inflect}
\usage{
inflect(x, threshold = 1)
}
\arguments{
\item{x}{The vector to search.}

\item{threshold}{The sensitivity; default is 1.}
}
\value{
Returns a list containing the minima and maxima.
}
\description{
Finds local minima and maxima in a vector.
}
\examples{
# ADD_EXAMPLES_HERE
}
