% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs2_wrapper.R
\name{macs2_callpeak}
\alias{macs2_callpeak}
\title{macs2 callpeak}
\usage{
macs2_callpeak(treated_bam_file, control_bam_file, experiment_name,
  output_folder, file_format = "AUTO", genome = "mm", pvalue = NULL,
  qvalue = 0.05, nomodel = T, extsize = NULL, run_as_sys = F)
}
\arguments{
\item{treated_bam_file}{DESCRIPTION.}

\item{control_bam_file}{DESCRIPTION.}

\item{experiment_name}{DESCRIPTION.}

\item{output_folder}{DESCRIPTION.}

\item{file_format}{DESCRIPTION.}

\item{genome}{DESCRIPTION.}

\item{pvalue}{DESCRIPTION.}

\item{qvalue}{DESCRIPTION.}

\item{nomodel}{DESCRIPTION.}

\item{extsize}{DESCRIPTION.}

\item{run_as_sys}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Wrapper for macs2 peak calling
}
\examples{
# ADD_EXAMPLES_HERE
}
