% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacktk_wrapper.R
\name{make_cut_matrix_atactk}
\alias{make_cut_matrix_atactk}
\title{Make Cut Matrix}
\usage{
make_cut_matrix_atactk(aggregate.setting = T, discrete.setting = F,
  bins = c("1-100", "180-1000"), resolution = 1, bam_file, bed_file,
  output_file)
}
\arguments{
\item{aggregate.setting}{DESCRIPTION.}

\item{discrete.setting}{DESCRIPTION.}

\item{bins}{DESCRIPTION.}

\item{resolution}{DESCRIPTION.}

\item{bam_file}{DESCRIPTION.}

\item{bed_file}{DESCRIPTION.}

\item{output_file}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Basic Wrapper for ATACTK: Makes a cut matrix based on the input criteria; currently returns a string with the command to run in terminal.
}
\examples{
# ADD_EXAMPLES_HERE
}
