% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixStats.R
\name{matrixPairwiseTest}
\alias{matrixPairwiseTest}
\title{Row level statistics from the pairwise comparison of matrix columns}
\usage{
matrixPairwiseTest(
  x,
  sampInfoDf,
  groupCol,
  numerator,
  denominator,
  metricFun = function(x, y) mean(x, na.rm = TRUE)/mean(y, na.rm = TRUE),
  statisticFun = function(x, y) nsTestT(x = x, y = y),
  pValueFun = function(x, y) nsTestPValue(x = x, y = y),
  adjustMethod = "BH",
  idName = "id",
  metricName = "logFc",
  statName = "t",
  pName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{x}{Input matrix.}

\item{sampInfoDf}{Column information data frame. Rownames of this data frame should match matrix columns.}

\item{groupCol}{Column from the sampInfoDf that indicates groups of columns (samples?).}

\item{numerator}{Group of columns used as the numerator (first element) of functions.}

\item{metricFun}{Function to obtain metric.}

\item{statisticFun}{Function to obtain statistic.}

\item{pValueFun}{Function to obtain p-value.}

\item{adjustMethod}{Method to perform p value adjustment.}

\item{idName}{Name for the id column, which is created using \link[base]{rownames}.}

\item{metricName}{Name for the metric column.}

\item{statName}{Name for the statistic column.}

\item{pName}{Name for the p value column.}

\item{pAdjName}{Name for the adjusted p value column.}

\item{denominato}{rGroup of columns used as the denominator (second element) of functions.}
}
\value{
A tidy data frame with results.
}
\description{
Divides the matrix into numerator and denominator using a column information data frame (sampInfoDf)
and applies several functions at row level to obtain a tidy data frame containing the results. It uses three functions to obtain:
1: a metric, which defaults to \link[stat]{mean} division, 2: a statistic, which defaults to
\link[biokit]{nsTestT} and 3: a p value, which defaults to \link[stats]{nsTestPValue}.
}
