% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixStats.R
\name{matrixSingleTest}
\alias{matrixSingleTest}
\title{Row level statistics of matrix}
\usage{
matrixSingleTest(
  x,
  metricFun = function(x) mean(x, na.rm = TRUE),
  statisticFun = function(x) nsTestT(x),
  pValueFun = function(x) nsTestPValue(x),
  adjustMethod = "BH",
  featName = "feature",
  metricName = "logFc",
  statName = "t",
  pName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{x}{Input matrix.}

\item{metricFun}{Function to obtain metric.}

\item{statisticFun}{Function to obtain statistic.}

\item{pValueFun}{Function to obtain p-value.}

\item{adjustMethod}{Method to perform p value adjustment.}

\item{featName}{Name for the feature id column, which is created using \link[base]{rownames}.}

\item{metricName}{Name for the metric column.}

\item{statName}{Name for the statistic column.}

\item{pName}{Name for the p value column.}

\item{pAdjName}{Name for the adjusted p value column.}
}
\value{
A tidy data frame with results.
}
\description{
Applies several functions over the rows of a matrix to obtain a tidy
data frame containing the results. It uses three functions to obtain: 1: a metric,
which defaults to \link[stat]{mean}, 2: a statistic, which defaults to
\link[biokit]{nsTestT} and 3: a p value, which defaults to \link[stats]{nsTestPValue}.
}
