% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixStats.R
\name{matrixTestFromContrasts}
\alias{matrixTestFromContrasts}
\title{Pairwise matrix test from contrasts}
\usage{
matrixTestFromContrasts(
  x,
  sampInfoDf,
  groupCol,
  contrasts,
  sep = "-",
  compName = "comparison",
  ...
)
}
\arguments{
\item{x}{Input matrix.}

\item{sampInfoDf}{Column information data frame. Rownames of this data frame should match matrix columns.}

\item{groupCol}{Column from the sampInfoDf that indicates groups of columns.}

\item{contrasts}{Vector containing the contrasts to analyze.}

\item{sep}{Character used to split contrasts into numerator and denominator.}

\item{compName}{Name for the column that indicates the pairwise comparison in the tidy data frame.}

\item{...}{Rest of arguments passed to \link[biokit]{matrixPairwiseTest}.}
}
\value{
A tidy data frame with results.
}
\description{
Performs the pairwise statistical analysis of the groups of columns of interest
using an input vector of contrasts.
}
