% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_molarity_related_functions.R
\name{Mcalc_grams}
\alias{Mcalc_grams}
\title{Calc grams for Molarity}
\usage{
Mcalc_grams(M_target, vol_target, mw, sentence = F, compoundID = NULL)
}
\arguments{
\item{M_target}{The molarity, given as M. Dbl.}

\item{vol_target}{The volume, given in liters. Dbl.}

\item{mw}{The molecular weight of the compound in question. Does not account for hydrated compounds. Dbl.}

\item{sentence}{Return the calculation as a description? Logical}

\item{compoundID}{If sentence == T, provide a label or the compound. String.}
}
\value{
Returns a string.
}
\description{
This function, part of a set of sub functions at the moment, calculates the grams required in a given volume to achieve a certain molarity.
}
\examples{
Mcalc_grams(M_target = 1,vol_target = 0.005,mw = 84.01,sentence = T,compoundID = "NaHCO~3~")
}
