% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_csv.R
\name{merge_csv_dir}
\alias{merge_csv_dir}
\title{Merge CSV Directory}
\usage{
merge_csv_dir(csv_directory, file.pattern = "*.csv", header = T,
  sep = ",", quote = "", stringsAsFactors = F, fill = T)
}
\arguments{
\item{csv_directory}{the object of the list.files function; full names suggested.}

\item{file.pattern}{a particular file object to search for. ex: "*.csv"}

\item{header}{do the files have a header?}

\item{sep}{what is the delimiter?}

\item{quote}{quote?}

\item{stringsAsFactors}{strings as factors?}

\item{fill}{fill NAs?}
}
\value{
Returns a df
}
\description{
Takes a list.files object, then reads in each object and binds them.  Requires objects to be the same structure.
}
\examples{
# df <- merge_csv_dir(csv_directory = "./csv")
}
