% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateMatrix.R
\name{messageMappingInfo}
\alias{messageMappingInfo}
\title{Evaluate and print translator df information}
\usage{
messageMappingInfo(df, sourceKey, targetKey)
}
\arguments{
\item{df}{A 2-columns translator data frame with source to target ids.}

\item{sourceKey}{Character indicating the column of the df where the source ids are stored.}

\item{targetKey}{Character indicating the column of the df where the source ids are stored.}
}
\description{
Evaluate and print translator df information
}
