% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gel_ladder.R
\name{model_gel_ladder}
\alias{model_gel_ladder}
\title{Model Gel Ladder}
\usage{
model_gel_ladder(ladder_bp_size, distance_migrated, polynomial_number = 2)
}
\arguments{
\item{ladder_bp_size}{size of the band. (100, 200, etc)}

\item{distance_migrated}{distance measurement for how far the band moved.}

\item{polynomial_number}{what polynomial to use for the model? default as 2.}
}
\value{
RETURN_DESCRIPTION
}
\description{
provides a model to predict base pair sizes from a ladder.
}
\examples{
# ADD_EXAMPLES_HERE
}
