% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{nsTest}
\alias{nsTest}
\title{Not sensitive T.Test}
\usage{
nsTest(...)
}
\arguments{
\item{...}{Arguments for \code{t.test} function.}
}
\value{
The P.value resulting from the T-Test or \code{NA} when any problem appears.
}
\description{
This function applies the T-Test function returning NA
instead of error when problems appear (as continuous). Additionally,
it extracts the P value from the resulting object and returns it when
available.
}
