% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{nsTestT}
\alias{nsTestT}
\title{Non sensitive T Test statistic}
\usage{
nsTestT(...)
}
\arguments{
\item{...}{Arguments passed to \link[stats]{t.test}.}
}
\value{
Statistic resulting from the T Test or \code{NA} when any problem appears.
}
\description{
Non sensitive T Test statistic
}
