% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraGsea.R
\name{oraFromList}
\alias{oraFromList}
\title{Over-representation analysis over a list of features}
\usage{
oraFromList(x, funCategories, minFeatures = 10, ...)
}
\arguments{
\item{x}{The list of features to be analysed.}

\item{funCategories}{Two-columns data frame containing the functional categories info.
First column should correspond to categories and the second to features.
Passed to \link[clusterProfiler]{enricher} TERM2GENE argument.}

\item{minFeatures}{Minimum features to evaluate.
The function will filter all list elements with a lower number of features than this value.}

\item{...}{Rest of arguments passed to \link[clusterProfiler]{enricher}}
}
\value{
A list of \link[clusterProfiler]{enricher} results, one per list element.
}
\description{
Uses a over-representation strategy (Fisher Exact T-Test) to
evaluate the over-represented categories in a list of features. Uses
the \link[clusterProfiler]{enricher} function to perform the analyses.
}
