% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{oraFromStats}
\alias{oraFromStats}
\title{Over representation analysis from statistical results}
\usage{
oraFromStats(
  df,
  funCatList,
  statusCol,
  noChangeLabel = "No change",
  featCol = "feature",
  ...
)
}
\arguments{
\item{df}{Data frame with results to analyze.}

\item{funCatList}{A list of character vectors with the functional categories to analyze.}

\item{statusCol}{Column containing the annotation of features.}

\item{noChangeLabel}{Label used to indicate not changing features. Rows with a statusCol value equal to this will be removed.}

\item{featCol}{Column containing the feature to submit for the over-representation analysis (i.e gene symbol).}

\item{...}{Other parameters passed to \link[biokit]{Over-representation analysis}.}
}
\value{
A tidy data frame with the results.
}
\description{
Performs an over-representation analysis from a data frame containing the results of
the statistical analysis of omic data. Uses the annotation column to split
the features and perform the over-representation analysis.
}
