% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{osTestMatrix}
\alias{osTestMatrix}
\title{One sample T-Test over log ratio matrix}
\usage{
osTestMatrix(
  x,
  adjustMethod = "BH",
  idName = "id",
  fcName = "logFc",
  pName = "pValue",
  pAdjName = "pAdj",
  ...
)
}
\arguments{
\item{x}{logFC matrix.}

\item{adjustMethod}{P value adjustment method.}

\item{idName}{Name for the resulting feature id column.}

\item{fcName}{Name for the resulting fold change column.}

\item{pName}{Name for the resulting p value column.}

\item{pAdjName}{Name for the resulting adjusted p value column.}

\item{...}{Rest of arguments passed to \link[stats]{t.test} function.}
}
\value{
A data frame with:
\item{id}{From row names.}
\item{logFC}{From row means.}
\item{pValue}{From the One Sample T-Test.}
\item{pAdj}{From multiple testing correction.}
}
\description{
Applies One Sample T-Test at row level in a matrix containing log transformed fold changes.
}
