% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_representation_analysis.R
\name{overRepresentationAnalysis}
\alias{overRepresentationAnalysis}
\title{Over-representation analysis}
\usage{
overRepresentationAnalysis(
  features,
  funCatList,
  universe,
  minSize = 1,
  maxSize = Inf,
  removeNoOverlap = TRUE,
  pAdjustMethod = "BH"
)
}
\arguments{
\item{features}{A vecor of characters that contains the features to analyze.}

\item{funCatList}{A list of character vectors with the functional categories to analyze.}

\item{universe}{A vector of characters to be used as the statistical background for the different tests. If not supplied, defaults to all the features in funCatList.}

\item{minSize}{Minimal size of a functional category to test. All functional categories below the threshold are excluded.}

\item{maxSize}{Minimal size of a functional category to test. All functional categories aboce the threshold are excluded.}

\item{removeNoOverlap}{Remove functional categories without overlapping features before performing the P adjustment?}

\item{pAdjustMethod}{The method used to correct estimated p values. Passed to \link[stats]{p.adjust}.}
}
\value{
A data frame with results in a tidy format
}
\description{
Uses two different approaches to assess the significance of the over-representation
of a feature vector over a list of functional categories. Uses \link[biokit]{fisherExactTest}.
}
