% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContrasts.R
\name{pairwiseContrasts}
\alias{pairwiseContrasts}
\title{Create pariwise contrasts}
\usage{
pairwiseContrasts(x, sep = "-")
}
\arguments{
\item{x}{Vector to be combined in pairs.}

\item{sep}{Separator for the resulting contrasts.}
}
\value{
A character vector containing all the possible pairwise contrasts.
}
\description{
Creates all the possible pairwise contrasts between
the unique elements of a vector.
}
