% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pairwise_comparison.R
\name{pairwiseMatrixTest}
\alias{pairwiseMatrixTest}
\title{Generic pairwise matrix test}
\usage{
pairwiseMatrixTest(
  mat,
  sampInfo,
  groupCol,
  numerator,
  denominator,
  metricFun = defaultLogFc,
  statFun = nsTestT,
  sigFun = nsTestPValue,
  pAdjustMethod = "BH",
  allowNa = TRUE,
  featName = "feature",
  metricName = "logFc",
  statName = "t",
  pValName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{mat}{Matrix containing data to be analyzed.}

\item{sampInfo}{Sample information data frame.}

\item{groupCol}{Name of the sampInfo column containing the grouping variable.}

\item{numerator}{Label of the grouping variable to be used as numerator.}

\item{denominator}{Label of the grouping variable to be used as denominator.}

\item{metricFun}{Function to calculate metric of change. Defaults to log2(ratio of means).}

\item{statFun}{Function to calculate statistic of change. Defaults to a non-sensitive T-Test T value.}

\item{sigFun}{Function to calculate significance of change. Defaults to a non-sensitive T-Test P value.}

\item{pAdjustMethod}{Method used to adjust P values.}

\item{allowNa}{Allow NAs on input matrix?}

\item{featName}{Name of the column containing the compared features (rows).}

\item{metricName}{Name of the column containing the metric.}

\item{statName}{Name of the column containing the statistic.}

\item{pValName}{Name of the column containing the p values.}

\item{pAdjName}{Name of the column containing the adjusted p values.}
}
\value{
A data frame with the row-wise statistical results.
}
\description{
Compare two pieces of data from a matrix given a
sample information data frame and a grouping column.
}
