% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{pcaPlot}
\alias{pcaPlot}
\title{Simple PCA analysis}
\usage{
pcaPlot(mat, sampInfo, groupCol)
}
\arguments{
\item{mat}{The input matrix containing normalized data or counts}

\item{sampInfo}{The sample information data frame}

\item{groupCol}{The column name of the grouping variable}
}
\value{
A ggplot2 object containing the resulting plot
}
\description{
Performs a principal component analysis of a
normalized expression/abundance matrix and plot the samples
in the bidimensional space formed by the two first principal
components. Samples are colored according to the variable of
interest in the sample information data frame.
}
