% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number_predict_confluency.R
\name{predict_cell_number}
\alias{predict_cell_number}
\title{Predict Cell Number}
\usage{
predict_cell_number(seeding_density, doubling_time, incubation_hours,
  return_text_statement = F)
}
\arguments{
\item{seeding_density}{DESCRIPTION.}

\item{doubling_time}{DESCRIPTION.}

\item{incubation_hours}{DESCRIPTION.}

\item{return_text_statement}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Predicts number of cells in culture based on seeding density, known doubling time information, and hours of incubation.
}
\examples{
# ADD_EXAMPLES_HERE
}
