% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.qubit.R
\name{read.qubit_csv}
\alias{read.qubit_csv}
\title{Read Qubit CSV}
\usage{
read.qubit_csv(file, header = T, sep = ",", quote = "", skip = 5)
}
\arguments{
\item{file}{File location. String}

\item{header}{Header? logical. T by default.}

\item{sep}{delimiter for import; String. "," by default}

\item{quote}{quote? String. "" by default.}

\item{skip}{How many rows to skip before reading.  Integer. 5 by default.}
}
\value{
Returns a data frame.
}
\description{
Read Qubit Results file.
}
\examples{
#read.qubit_csv(file = "data-raw/Qubit2018-05-30_14-20-15_ChIP_CTCF_Pulldown.csv")
}
