% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarsCovMat}
\alias{sarsCovMat}
\title{RNA-Seq data of human cell lines infected with sars-cov-2}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 21797 rows and 12 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE147507}
}
\usage{
sarsCovMat
}
\description{
A matrix containing the gene-level counts for the GEO entry GSE147507.
}
\keyword{datasets}
