% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarsCovSampInfo}
\alias{sarsCovSampInfo}
\title{Sample level information for the sars-cov-2 count matrix}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE147507}
}
\usage{
sarsCovSampInfo
}
\description{
A data frame containing the relevant information for each sample of the GSE147507 study.
}
\keyword{datasets}
