% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMatrixTest.R
\name{singleMatrixTest}
\alias{singleMatrixTest}
\title{Row level statistics of matrix}
\usage{
singleMatrixTest(
  x,
  metricFun = function(x) mean(x, na.rm = TRUE),
  statisticFun = function(x) nsTestT(x),
  pValueFun = function(x) nsTestPValue(x),
  adjustMethod = "BH",
  idName = "id",
  metricName = "logFc",
  statName = "t",
  pName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{x}{Matrix to analyze.}

\item{metricFun}{Function to obtain the metric.}

\item{statisticFun}{Function to obtain the statistic.}

\item{pValueFun}{Function to obtain significance as p-value.}

\item{adjustMethod}{Method to perform p value adjustment.}

\item{idName}{Name for the id column, which is created using \link[base]{rownames}.}

\item{metricName}{Name for the metric column.}

\item{statName}{Name for the statistic column.}

\item{pName}{Name for the p value column.}

\item{pAdjName}{Name for the adjusted p value column.}
}
\value{
A tidy data frame with results.
}
\description{
Applies several functions over the rows of a matrix to obtain a tidy
data frame containing the results. Defaults to a One-Sample T-Test over
the input matrix
}
