% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer_functions.R
\name{split_motifs}
\alias{split_motifs}
\title{FUNCTION_TITLE}
\usage{
split_motifs(annotated_df, motif_col, zero_based = F, filter_for_motifs = F,
  motif_label = NULL, return_bedfile_format = F)
}
\arguments{
\item{annotated_df}{DESCRIPTION.}

\item{motif_col}{DESCRIPTION.}

\item{zero_based}{DESCRIPTION.}

\item{filter_for_motifs}{DESCRIPTION.}

\item{motif_label}{DESCRIPTION.}

\item{return_bedfile_format}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE

}
