% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsToFunc.R
\name{splitFeatures}
\alias{splitFeatures}
\title{Split data frame into list of features}
\usage{
splitFeatures(
  x,
  annotCol = "status",
  featCol = "feature",
  labelToRemove = "No change",
  splitCol = NULL,
  sep = "_"
)
}
\arguments{
\item{x}{Data frame to split.}

\item{annotCol}{Column containing the annotation label.}

\item{featCol}{Column containing the feature id. Coerced to character.}

\item{labelToRemove}{Label assigned to the elements which will not be splitted.}

\item{splitCol}{Additional column to use for splitting.}

\item{sep}{Separator of the new splitting variable.}
}
\value{
A list of character vectors.
}
\description{
Uses annotation labels to split a data frame of results into
a list containing a character vector of features. If an additional split column is provided,
then it split by the combination of the annotation and the split column.
}
