% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{splitFunMerge}
\alias{splitFunMerge}
\title{Split, Function and Merge}
\usage{
splitFunMerge(df, splitCol, fun)
}
\arguments{
\item{df}{Data frame to process.}

\item{splitCol}{Column to split by.}

\item{fun}{Function to be applied to each sub-data frame.}
}
\value{
The processed data frame.
}
\description{
Splits the input data frame using a grouping column, applies
the provided function over each list element and return the row-binded
resulting data frame.
}
