% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioUtils.R
\name{summariseMatrix}
\alias{summariseMatrix}
\title{Translate input matrix using a translator df}
\usage{
summariseMatrix(x, df, sourceKey, targetKey, summariseFun)
}
\arguments{
\item{x}{The input matrix. Should have rownames.}

\item{df}{A 2-columns translator data frame with source to target ids.}

\item{sourceKey}{Character indicating the column of the df where the source ids are stored.}

\item{targetKey}{Character indicating the column of the df where the source ids are stored.}

\item{summariseFun}{Function used to resolve multi-mapping situations.}
}
\value{
A matrix with translated and summarised ids.
}
\description{
Translates the rownames of the input matrix
into the desired ids. When input ids maps to several
target ids, uses the input function to resolve conflicts.
}
