% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox_in_vitro_planner.R
\name{tox_in_vitro_planner}
\alias{tox_in_vitro_planner}
\title{In vitro Tox Experiment Planner}
\usage{
tox_in_vitro_planner(cell.type, treatment, time, bio_rep_number)
}
\arguments{
\item{cell.type}{Vector of strings containing the cell types.}

\item{treatment}{vector of strings containing the treatments.}

\item{time}{Vector of strings containing the treatment times.}

\item{bio_rep_number}{integer, how many biological replicates?}
}
\value{
returns a dataframe.
}
\description{
Estimator for planning the number of plates and the conditions required, based on what is provided.
}
\examples{
# tox_in_vitro_planner(cell.type = c("HeLa","Other"),treatment = c("DMSO","compound A","compound B"),time = c("2 hours","4 hours"),bio_rep_number = 3)
}
