% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_matrix.R
\name{translateMatrix}
\alias{translateMatrix}
\title{Translate matrix rownames}
\usage{
translateMatrix(mat, df, sourceKey, targetKey, summariseFun)
}
\arguments{
\item{mat}{Input matrix. Should have rownames.}

\item{df}{A 2-columns translator data frame with source to target ids.}

\item{sourceKey}{Character indicating the column of the df where the source ids are stored.}

\item{targetKey}{Character indicating the column of the df where the source ids are stored.}

\item{summariseFun}{Function used to resolve multi-mapping situations.}
}
\value{
A matrix with translated and summarised ids.
}
\description{
Translates the rownames of the input matrix into the
desired ids using a translator data frame. When
input ids maps to several target ids, uses the summarise
function to resolve conflicts.
}
