% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_matrix.R
\name{translateMatrixWithDb}
\alias{translateMatrixWithDb}
\title{Translate matrix rownames using annotation package}
\usage{
translateMatrixWithDb(mat, db, sourceKey, targetKey, summariseFun)
}
\arguments{
\item{mat}{Input matrix.}

\item{db}{Annotation package object to use for the translation.}

\item{sourceKey}{Source key in the DB.}

\item{targetKey}{Target key in the DB.}

\item{summariseFun}{Function used to resolve multi-mapping situations.}
}
\value{
A matrix with translated and summarised ids.
}
\description{
Uses an annotation package like
\href{https://bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html}{org.Hs.eg.db} or
\href{https://bioconductor.org/packages/release/data/annotation/html/org.Mm.eg.db.html}{org.Mm.eg.db}
to translate the rownames of the input matrix using \link[biokit]{translateMatrix}.
}
