% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrappers.R
\name{translateSE}
\alias{translateSE}
\title{Translate SummarizedExperiment}
\usage{
translateSE(se, df, sourceKey, targetKey, summariseFun)
}
\arguments{
\item{se}{The SummarizedExperiment object that contains the matrix to be translated.}

\item{df}{A 2-columns translator data frame with source to target ids.}

\item{sourceKey}{Character indicating the column of the df where the source ids are stored.}

\item{targetKey}{Character indicating the column of the df where the source ids are stored.}

\item{summariseFun}{Function used to resolve multi-mapping situations.}
}
\value{
The translated version of the summarized experiment.
}
\description{
Translates the first assay of a SummarizedExperiment.
}
