% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrappers.R
\name{translateSEWithDb}
\alias{translateSEWithDb}
\title{Translate SummarizedExperiment with annotation package}
\usage{
translateSEWithDb(se, db, sourceKey, targetKey, summariseFun)
}
\arguments{
\item{se}{The SummarizedExperiment object that contains the matrix to be translated.}

\item{db}{Annotation package object to use for the translation.}

\item{sourceKey}{Source key in the DB.}

\item{targetKey}{Target key in the DB.}

\item{summariseFun}{Function used to resolve multi-mapping situations.}
}
\value{
The translated version of the summarized experiment.
}
\description{
Translate SummarizedExperiment with annotation package
}
