% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_matrix.R
\name{validateMatrix}
\alias{validateMatrix}
\title{Validate matrix}
\usage{
validateMatrix(mat)
}
\arguments{
\item{mat}{The matrix to validate.}
}
\value{
TRUE if no errors are found.
}
\description{
Evaluates integrity of the input matrix.
}
