% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateSampInfo}
\alias{validateSampInfo}
\title{Validate sample information}
\usage{
validateSampInfo(
  sampInfo,
  groupCol,
  mat,
  checkNames = TRUE,
  checkSingleSample = TRUE
)
}
\arguments{
\item{sampInfo}{The sample information data frame.}

\item{groupCol}{The column containing the grouping variable.}

\item{mat}{Matrix containing data. If provided, it will check number of columns in mat.}

\item{checkNames}{Check if grouping variable contains syntactically valid names? If TRUE, they are fixed in the output data frame.}

\item{checkSingleSample}{Check if any group contains only one sample? A warning message appears if true.}
}
\value{
The checked sampInfo data frame
}
\description{
Evaluates integrity of the grouping variable at the sample information data frame.
}
