% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_frame.R
\name{validateTranslatorDf}
\alias{validateTranslatorDf}
\title{Validate translator data frame}
\usage{
validateTranslatorDf(df, sourceKey, targetKey)
}
\arguments{
\item{df}{The translator data frame to validate.}
}
\value{
The validated and ammended translator
data frame if no problems are found.
}
\description{
Evaluates integrity of the translator data frame.
}
