% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{violinPlot}
\alias{violinPlot}
\title{Violin plot}
\usage{
violinPlot(mat)
}
\arguments{
\item{mat}{The matrix to be plotted}
}
\value{
A ggplot2 object
}
\description{
Creates a per-column violin plot to explore data density before and after normalization
}
