% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Simple Volcano Plot}
\usage{
volcanoPlot(intTable, logFcCutoff = log2(2), pCutoff = 0.05)
}
\arguments{
\item{intTable}{Data frame with differential analysis results}

\item{logFcCutoff}{Fold change cutoff}

\item{pCutoff}{Adjusted P value cutoff}
}
\value{
A ggplot2 object containing the resulting plot
}
\description{
Creates a simple volcano plot summarizing the number of differentially altered
features by comparison.
}
