% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{annotateByCutoff}
\alias{annotateByCutoff}
\title{Add status column to comparison data frame using cutoff}
\usage{
annotateByCutoff(
  x,
  fcCutoff = 1,
  pCutoff = 0.05,
  splitUpDown = TRUE,
  fcCol = "logFc",
  pCol = "pAdj",
  noChangeLabel = "No change",
  statusCol = "status",
  upLabel = "Up",
  downLabel = "Down",
  sigLabel = "Significant"
)
}
\arguments{
\item{x}{Data frame to annotate.}

\item{fcCutoff}{Fold change cutoff. Should be positive, as it is transformed into negative when annotating 'downregulated' features.}

\item{pCutoff}{P Value cutoff.}

\item{splitUpDown}{Wethter to split generated label into up and down features.}

\item{fcCol}{Column used to filter by fc.}

\item{pCol}{Column used to filter by p value.}

\item{noChangeLabel}{Label used for not-changing features.}

\item{statusCol}{Name of the new column containing labels.}

\item{upLabel}{Label for up-regulated features. Only used if \code{splitUpDown = TRUE}.}

\item{downLabel}{Label for down-regulated features. Only used if \code{splitUpDown = TRUE}.}

\item{sigLabel}{Label for significantly altered features. Only used if \code{splitUpDown = FALSE}.}
}
\value{
Annotated data frame.
}
\description{
Uses a fold change and P value cutoff to annotate a
data frame containing the results of a stat comparison.
}
