% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{annotateTopN}
\alias{annotateTopN}
\title{Add status column to comparison data frame using top N features}
\usage{
annotateTopN(
  x,
  n,
  sortCol,
  splitUpDown = TRUE,
  decreasing = FALSE,
  twoSides = FALSE,
  fcCol = "logFc",
  noChangeLabel = "No change",
  statusCol = "status",
  upLabel = "Up",
  downLabel = "Down",
  sigLabel = "Significant"
)
}
\arguments{
\item{x}{Data frame to annotate.}

\item{n}{Top N features to be annotated}

\item{sortCol}{Column used to sort the data frame.}

\item{splitUpDown}{Wether to split top N features into up and down regulated using a fold change column. This will annotate 2*n features.}

\item{decreasing}{Order sortCol decreasing.}

\item{twoSides}{Pick up both extremes of the sorted data frame as the top N features. SplitUpDown should be TRUE.}

\item{fcCol}{Fold change column used to sepparate up and down features.}

\item{noChangeLabel}{Label used for not-changing features.}

\item{statusCol}{Name of the new column containing labels.}

\item{upLabel}{Label for up-regulated features. Only used if \code{splitUpDown = TRUE}.}

\item{downLabel}{Label for down-regulated features. Only used if \code{splitUpDown = TRUE}.}

\item{sigLabel}{Label for significantly altered features. Only used if \code{splitUpDown = FALSE}.}
}
\value{
Annotated data frame.
}
\description{
Orders a data frame and adds a column to indicate the
status of the top N features.
}
