% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaUtils.R
\name{autoLimma}
\alias{autoLimma}
\title{Automatic limma analysis from SummarizedExperiment}
\usage{
autoLimma(se, groupColumn, ...)
}
\arguments{
\item{se}{The SummarizedExperiment object to analyze.}

\item{groupColumn}{The id of the \link[SummarizedExperiment]{colData} column that is used to create the pairwise contrasts.}

\item{...}{Arguments passed to the \link[biokit]{limmaDfFromContrasts} function.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Uses data stored in a SummarizedExperiment object to
automatically perform all the possible pairwise comparisons
between the groups defined in a \link[SummarizedExperiment]{colData} column.
}
