% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{contrastsFromDesign}
\alias{contrastsFromDesign}
\title{Create contrast matrix given a design matrix.}
\usage{
contrastsFromDesign(x)
}
\arguments{
\item{x}{The design matrix.}
}
\value{
A contrast matrix with all the possible pairwise comparisons.
}
\description{
Uses levels defined in the column names of the design matrix
to create a contrast matrix with \link[limma]{makeContrasts}.
}
