% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDesign.R
\name{designFromSampInfo}
\alias{designFromSampInfo}
\title{Create design matrix from sample metadata}
\usage{
designFromSampInfo(x, column)
}
\arguments{
\item{x}{The data frame containing the column metadata (e.g sample grouping).}

\item{column}{The data frame column that is used to compose the design matrix.}
}
\value{
The composed design matrix.
}
\description{
Uses a data frame containing sample metadata to create a design matrix
for subsequent analyses.
}
