% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTools2_wrapper.R
\name{dT2_computeMatrix}
\alias{dT2_computeMatrix}
\title{dT2 computeMatrix}
\usage{
dT2_computeMatrix(bigwig_files, bed_files, scale_regions = F,
  ref_point = "TSS", bin_size = 10, upstream_bp = 1, downstream_bp = 1,
  missingDataAsZero = T, skipZeros = T, blacklist_file = NULL, output_gz,
  outFileMatrix = NULL, outSortedRegions = NULL, run_as_sys = F)
}
\arguments{
\item{bigwig_files}{DESCRIPTION.}

\item{bed_files}{DESCRIPTION.}

\item{scale_regions}{DESCRIPTION.}

\item{ref_point}{DESCRIPTION.}

\item{bin_size}{DESCRIPTION.}

\item{upstream_bp}{DESCRIPTION.}

\item{downstream_bp}{DESCRIPTION.}

\item{missingDataAsZero}{DESCRIPTION.}

\item{skipZeros}{DESCRIPTION.}

\item{blacklist_file}{DESCRIPTION.}

\item{output_gz}{DESCRIPTION.}

\item{outFileMatrix}{DESCRIPTION.}

\item{outSortedRegions}{DESCRIPTION.}

\item{run_as_sys}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Wrapper for deepTools2 computeMatrix
}
\examples{
# ADD_EXAMPLES_HERE
}
