% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTools2_wrapper.R
\name{dT2_plotHeatmap}
\alias{dT2_plotHeatmap}
\title{dT2 plot Heatmap}
\usage{
dT2_plotHeatmap(gz_file, output_file, colorMap = "viridis", dpi = 600,
  run_as_sys = F, outFileSortedRegions = NULL, outFileNameMatrix = NULL)
}
\arguments{
\item{gz_file}{DESCRIPTION.}

\item{output_file}{DESCRIPTION.}

\item{colorMap}{DESCRIPTION.}

\item{dpi}{DESCRIPTION.}

\item{run_as_sys}{DESCRIPTION.}

\item{outFileSortedRegions}{DESCRIPTION.}

\item{outFileNameMatrix}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Wrapper for deepTools2 plotHeatmap command.
}
\examples{
# ADD_EXAMPLES_HERE
}
