% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanHallmarks}
\alias{humanHallmarks}
\title{Human MSigDb hallmarks in TERM2GENE format}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7322 rows and 2 columns.}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp}
}
\usage{
humanHallmarks
}
\description{
Human MSigDb hallmarks in TERM2GENE format
}
\keyword{datasets}
