% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{limmaDfFromContrasts}
\alias{limmaDfFromContrasts}
\title{Obtain limma results from data, design and contrasts matrix.}
\usage{
limmaDfFromContrasts(
  x,
  desMat,
  conMat,
  compName = "comparison",
  featName = "feature",
  fcName = "logFc",
  pName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{x}{Matrix with data to analyse.}

\item{desMat}{Design matrix.}

\item{conMat}{Contrasts matrix.}

\item{compName}{Name for the column that indicates the pairwise comparison in the tidy data frame.}

\item{featName}{Name for the column that indicates the analyzed row feature the tidy data frame.}

\item{fcName}{Name for the resulting fold change column.}

\item{pName}{Name for the resulting p value column.}

\item{pAdjName}{Name for the resulting adjusted p value column.}
}
\value{
A data frame with the results in a tidy format.
}
\description{
Creates a linear model with limma and fits it to the provided contrasts. Then uses \code{eBayes()}
to smooth standard errors and obtain the list of pairwise comparison results with \code{topTable()}.
}
