% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{nsTest}
\alias{nsTest}
\title{Not sensitive T Test}
\usage{
nsTest(...)
}
\arguments{
\item{...}{Arguments for the \link[stats]{t.test} function.}
}
\value{
P value resulting from the Test or \code{NA} when any problem appears.
}
\description{
Applies the T Test function returning NA
instead of error when problems appear. Additionally,
it extracts the P value from the resulting object when available.
}
