% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{nsTestPValue}
\alias{nsTestPValue}
\title{Not sensitive T test p value}
\usage{
nsTestPValue(...)
}
\arguments{
\item{...}{Arguments passed to \link[stats]{t.test}.}
}
\value{
P value resulting from the T Test or \code{NA} when any problem appears.
}
\description{
Not sensitive T test p value
}
