% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{pcaToList}
\alias{pcaToList}
\title{Perform PCA and create list with results}
\usage{
pcaToList(x, transpose = TRUE, roundDigits = 2, ...)
}
\arguments{
\item{x}{The matrix to analyze.}

\item{transpose}{Wether to transpose or not the input matrix.}

\item{roundDigits}{Number of digits for the rounded percent of
variance explained by each PC.}

\item{...}{Other parameters passed to the \link[stats]{prcomp} function.}
}
\value{
A list with three elements:
\item{result}{The prcomp object containing the results of the PCA.}
\item{summary}{The summary objects containing the matrix with PC values.}
\item{pcts}{A character vector with the variance explained by each PC.}
}
\description{
Performs principal component analysis for a given matrix
and returns a list with the formatted results.
}
