% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_individual_analysis.R
\name{singleMatrixTest}
\alias{singleMatrixTest}
\title{Generic single matrix analysis}
\usage{
singleMatrixTest(
  mat,
  metricFun = function(x) mean(x, na.rm = TRUE),
  statFun = nsTestT,
  pValueFun = nsTestPValue,
  allowNa = TRUE,
  pAdjustMethod = "BH",
  featName = "feature",
  metricName = "logFc",
  statName = "t",
  pValName = "pValue",
  pAdjName = "pAdj"
)
}
\arguments{
\item{mat}{Matrix containing data to be analyzed.}

\item{metricFun}{Function to calculate metric of change. Defaults to log2(ratio of means).}

\item{statFun}{Function to calculate statistic of change. Defaults to a non-sensitive T-Test T value.}

\item{allowNa}{Allow NAs on input matrix?}

\item{pAdjustMethod}{Method used to adjust P values.}

\item{featName}{Name of the column containing the compared features (rows).}

\item{metricName}{Name of the column containing the metric.}

\item{statName}{Name of the column containing the statistic.}

\item{pValName}{Name of the column containing the p values.}

\item{pAdjName}{Name of the column containing the adjusted p values.}

\item{sigFun}{Function to calculate significance of change. Defaults to a non-sensitive T-Test P value.}
}
\value{
A data frame with the row-wise statistical results.
}
\description{
Evaluates matrix data row-wise. This function is useful when the matrix
consists on relative measures (i.e a matrix of log2 fold changes) or to analyze
other features such as normal distribution.
}
