% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioUtils.R
\name{translateMatrix}
\alias{translateMatrix}
\title{Translate the input matrix rownames to the desired DB ids}
\usage{
translateMatrix(x, db, sourceKey, targetKey, summariseFun)
}
\arguments{
\item{x}{The input matrix. Should have rownames.}

\item{db}{The database to use for the translation.}

\item{sourceKey}{The source key in the DB.}

\item{targetKey}{The target key in the DB.}

\item{summariseFun}{Function used to resolve multi-mapping situations.}
}
\value{
A matrix with translated and summarised ids.
}
\description{
Uses a database like org.Hs.eg.db (https://bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html) or
org.Mm.eg.db (https://bioconductor.org/packages/release/data/annotation/html/org.Mm.eg.db.html) to translate the rownames of the input matrix
into the desired ids. When input ids maps to several target ids, uses the input function to resolve conflicts.
}
