% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateMatrix.R
\name{translateMatrix}
\alias{translateMatrix}
\title{Translate the input matrix rownames to the desired DB ids}
\usage{
translateMatrix(x, db, sourceKey, targetKey, summariseFun)
}
\arguments{
\item{x}{Input matrix. Should have rownames.}

\item{db}{Database to use for the translation.}

\item{sourceKey}{Source key in the DB.}

\item{targetKey}{Target key in the DB.}

\item{summariseFun}{Function used to resolve multi-mapping situations.}
}
\value{
A matrix with translated and summarised ids.
}
\description{
Uses a database like \href{https://bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html}{org.Hs.eg.db} or
\href{https://bioconductor.org/packages/release/data/annotation/html/org.Mm.eg.db.html}{org.Mm.eg.db}
to translate the rownames of the input matrix
into the desired ids. When input ids maps to several target ids, uses the input function to resolve conflicts.
}
