% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateMatrix}
\alias{validateMatrix}
\title{Validate matrix}
\usage{
validateMatrix(mat, checkRowNames = TRUE, allowNa = TRUE)
}
\arguments{
\item{mat}{The matrix to validate.}

\item{checkRowNames}{Check rownames are present? Stop execution if TRUE and mat does not contain rownames.}

\item{allowNa}{Allow NA presence in mat?  Stop execution if NAs are found in mat.}
}
\value{
TRUE if no errors are found.
}
\description{
Evaluates integrity of the input matrix.
}
