% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateTranslatorDf}
\alias{validateTranslatorDf}
\title{Validate translator data frame}
\usage{
validateTranslatorDf(df, sourceKey, targetKey)
}
\arguments{
\item{df}{The translator data frame to validate.}

\item{sourceKey}{Character indicating the column of the df where the source ids are stored.}

\item{targetKey}{Character indicating the column of the df where the source ids are stored.}
}
\value{
The validated and ammended translator
data frame if no problems are found.
}
\description{
Evaluates integrity of the translator data frame, subsetting it
to selected columns and removing duplicated rows.
}
